Security + een paar style changes

+++ database/api/benodigdheid.js
@@ -1,8 +1,8 @@
-import Database from "../database.js";
-const database = await Database.init();
+import {verifyAdmin} from "../../util/jwt-auth.js";
+
 const url = "/api/benodigdheid";
 
-export default function BenodigdheidAPI(app) {
+export default function BenodigdheidAPI(app, database) {
     app.get(url, async (req, res) => {
         if (req.query.id) {
             const benodigdheid = (await database.query("SELECT * FROM benodigdheid WHERE id = ?", [req.query.id]))[0];
@@ -14,6 +14,7 @@ export default function BenodigdheidAPI(app) {
     });
 
     app.post(url, async (req, res) => {
+        if (!(await verifyAdmin(req.query.token))) return res.status(401).send("Unauthorized");
         if (req.body.naam && req.body.activiteit) {
             const benodigdheid = await database.query("INSERT INTO benodigdheid (naam, activiteit) VALUES (?, ?)", [req.body.naam, req.body.activiteit]);
             res.send(benodigdheid);
@@ -23,6 +24,7 @@ export default function BenodigdheidAPI(app) {
     });
 
     app.put(url, async (req, res) => {
+        if (!(await verifyAdmin(req.query.token))) return res.status(401).send("Unauthorized");
         if (req.body.id && req.body.naam && req.body.activiteit) {
             const benodigdheid = await database.query("UPDATE benodigdheid SET naam = ?, activiteit = ? WHERE id = ?", [req.body.naam, req.body.activiteit, req.body.id]);
             res.send(benodigdheid);
@@ -32,6 +34,7 @@ export default function BenodigdheidAPI(app) {
     });
 
     app.delete(url, async (req, res) => {
+        if (!(await verifyAdmin(req.query.token))) return res.status(401).send("Unauthorized");
         if (req.query.id) {
             const benodigdheid = await database.query("DELETE FROM benodigdheid WHERE id = ?", [req.query.id]);
             res.send(benodigdheid);
